# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VentraPOS API",
    "version": "1.0.0",
    "description": "HTTP API for VentraPOS. Authenticated calls support `Authorization: Bearer <access_token>` (JSON from POST /api/auth/login when enabled) and/or the `__ventra_at` access cookie. Send `X-Branch-Id` to scope a branch. See docs/API-SECURITY.md."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local (pnpm dev) — use this in API reference to Try it"
    },
    {
      "url": "https://www.ventrapos.com",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access JWT"
      },
      "ventraAccessCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__ventra_at",
        "description": "HttpOnly access cookie (same-origin and Electron)."
      },
      "branchHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Branch-Id",
        "description": "Active branch id (optional)."
      },
      "paystackSignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Paystack-Signature"
      },
      "cronSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "CRON_SECRET in Authorization header"
      },
      "uploadthing": {
        "type": "apiKey",
        "in": "header",
        "name": "x-uploadthing-version"
      },
      "platformKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Ventra-Platform-Key",
        "description": "Platform (superadmin) key from VENTRA_PLATFORM_API_KEYS. Use with X-Act-As-Business-Id on tenant routes, or alone on /api/platform/*."
      },
      "actAsBusiness": {
...
  },
  "security": [],
  "paths": {
...
    "/api/platform/users": {
      "get": {
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformKey": []
          }
        ],
        "summary": "GET /api/platform/users",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformKey": []
          }
        ],
        "summary": "POST /api/platform/users",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/platform/users/{id}": {
      "get": {
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformKey": []
          }
        ],
        "summary": "GET /api/platform/users/{id}",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformKey": []
          }
        ],
        "summary": "PATCH /api/platform/users/{id}",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "platform"
        ],
        "security": [
          {
            "platformKey": []
          }
        ],
        "summary": "DELETE /api/platform/users/{id}",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  }
}